feat(PRLT-1365): consolidate Docker runners into docker-management.ts; delete docker.ts
